+++ src/blas-mxfp6gemm-cuda/helper.h
@@ -10,7 +10,8 @@
 #include <vector>
 
 #include <cublasLt.h>
-#include <cuda_fp16.h>
+//#include <cuda_fp16.h>
+#include <cuda_bf16.h>
 #include <cuda_fp8.h>
 #include <cuda_runtime_api.h>
 
@@ -35,7 +36,7 @@ inline void checkCublasStatus(cublasStatus_t status) {
 // A and B are FP6 (E2M3) packed four elements per three bytes along K.
 // Scales are E8M0 (UE8M0), one scale per 32-element block along K
 // (SCALE_BLOCK_SIZE = 32), matching the OCP microscaling (MX) format.
-// Output D is FP16.
+// Output D is BF16.
 struct Mxfp6TestBench {
     // SCALE_BLOCK_SIZE = 32 for mxfp6 (one UE8M0 scale per 32 FP6 elements along K)
     static constexpr int SCALE_BLOCK_SIZE = 32;
@@ -61,7 +62,7 @@ struct Mxfp6TestBench {
         checkCudaStatus(cudaMalloc(&Bdev, bBytes));
         checkCudaStatus(cudaMalloc(&AscaleDev, aScaleBytes));
         checkCudaStatus(cudaMalloc(&BscaleDev, bScaleBytes));
-        checkCudaStatus(cudaMalloc(reinterpret_cast<void**>(&Ddev), dElems * sizeof(__half)));
+        checkCudaStatus(cudaMalloc(reinterpret_cast<void**>(&Ddev), dElems * sizeof(__nv_bfloat16)));
         checkCudaStatus(cudaMalloc(&workspace, workspaceSize));
         checkCudaStatus(cudaStreamCreate(&stream));
 
@@ -140,8 +141,8 @@ struct Mxfp6TestBench {
     //   D = alpha * (A*a_scale) @ (B*b_scale)^T + beta * C
     // equals alpha * (a_elem*a_scale) * (b_elem*b_scale) * K (with beta = 0).
     bool verify(double relTol = 1e-2) {
-        std::vector<__half> Dh(dElems);
-        checkCudaStatus(cudaMemcpy(Dh.data(), Ddev, dElems * sizeof(__half), cudaMemcpyDeviceToHost));
+        std::vector<__nv_bfloat16> Dh(dElems);
+        checkCudaStatus(cudaMemcpy(Dh.data(), Ddev, dElems * sizeof(__nv_bfloat16), cudaMemcpyDeviceToHost));
 
         float aElem  = decodeE2M3(FP6_FILL);
         float bElem  = decodeE2M3(FP6_FILL);
@@ -152,7 +153,7 @@ struct Mxfp6TestBench {
 
         double maxErr = 0.0;
         for (size_t i = 0; i < dElems; i++) {
-            double got = (double)__half2float(Dh[i]);
+            double got = (double)__bfloat162float(Dh[i]);
             maxErr = std::max(maxErr, std::fabs(got - expected));
         }
         bool ok = maxErr <= relTol * (std::fabs(expected) + 1.0);
@@ -168,7 +169,7 @@ struct Mxfp6TestBench {
 
     void *Adev = nullptr, *Bdev = nullptr;           // packed FP6 (E2M3)
     void *AscaleDev = nullptr, *BscaleDev = nullptr;  // UE8M0 block scales
-    __half *Ddev = nullptr;                           // FP16 output
+    __nv_bfloat16 *Ddev = nullptr;                    // BF16 output
     void *workspace = nullptr;
     cudaStream_t stream;
     cublasLtHandle_t ltHandle;
