+++ src/blas-mxfp8gemm-sycl/main.cpp
@@ -10,11 +10,11 @@ using namespace dnnl;
 // A and B are FP8 (f8_e4m3), one byte per element along K.
 // The block scales are E8M0 (e8m0) with one scale per 32 elements along the K
 // dimension, expressed through per-argument scale attributes with group size 32
-// (matching the OCP microscaling format). Output D is FP16.
+// (matching the OCP microscaling format). Output D is BF16.
 //
 // This follows the oneDNN mxfp8 matmul example
 // (https://uxlfoundation.github.io/oneDNN/example_mxfp_matmul.cpp.html) but
-// keeps the FP16 output and the benchmark structure of blas-mxfp8gemm-cuda.
+// keeps the BF16 output and the benchmark structure of blas-mxfp8gemm-cuda.
 //
 // A is stored (M,K) row-major and B is stored (N,K) row-major. B is described
 // as {K,N} with strides {1,K} so that it is interpreted as B^T, producing
@@ -26,7 +26,7 @@ bool OnednnMxfp8Matmul(const int repeat, Mxfp8TestBench &tb) {
 
     auto a_md = memory::desc({m, k}, memory::data_type::f8_e4m3, memory::dims{k, 1});
     auto b_md = memory::desc({k, n}, memory::data_type::f8_e4m3, memory::dims{1, k});
-    auto c_md = memory::desc({m, n}, memory::data_type::f16,     memory::dims{n, 1});
+    auto c_md = memory::desc({m, n}, memory::data_type::bf16,    memory::dims{n, 1});
 
     // Per-argument E8M0 block scales along K (group size = SCALE_BLOCK_SIZE).
     // mask selects both dimensions so scales vary per row and per K-block.
